NodeGroupMap separate class
NodeGroupMap is now separate class; added code for node group ratios
